Move goal widget data to separate files
